Detect enum values for Enums in array schema
According to https://mongoosejs.com/docs/subdocs.html#altsyntax
a schema for an array of objects can be declared as a mongoose
schema instance or a plain object.

The type is generated using schemaType when it is a mongoose
schema instance, else objectType is used.

This commit adds the detection of enum values to schemaType
as this was only supported in objectType (when the schema
is defined as a plain object).

Signed-off-by: Chris Lahaye <dev@chrislahaye.com>