feat: extract a isGenerated property for each field, allow to create entry specifying the _id
A new property named isGenerated is added to fields' descriptions. It represents the fact that a value will be generated on document creation. When this value will be received as `false` by the frontend, it will allow to specify a value when creating a document even if the field is read only (otherwise it would be impossible to create documents from the UI).

Linked to CU-48py0a